Implemented Movt of Cars

+++ src/Position.java
@@ -1,18 +1,18 @@
 import java.util.Random;
 
 public class Position {
-    private int x;
-    private int y;
+    private float x;
+    private float y;
 
-    public int getX() {
+    public float getX() {
         return this.x;
     }
 
-    public int getY() {
+    public float getY() {
         return this.y;
     }
 
-    public Position(int x, int y) {
+    public Position(float x, float y) {
         this.x = x;
         this.y = y;
     }
